--- packages/thirdweb/src/extensions/prebuilts/deploy-erc20.test.ts
@@ -1,56 +1,0 @@
-import { describe, expect, it } from "vitest";
-import { ANVIL_CHAIN } from "../../../test/src/chains.js";
-import { TEST_CLIENT } from "../../../test/src/test-clients.js";
-import { TEST_ACCOUNT_D } from "../../../test/src/test-wallets.js";
-import { getContract } from "../../contract/contract.js";
-import { name } from "../common/read/name.js";
-import { deployERC20Contract } from "./deploy-erc20.js";
-
-const account = TEST_ACCOUNT_D;
-
-// skip this test suite if there is no secret key available to test with
-// TODO: remove reliance on secret key during unit tests entirely
-describe.runIf(process.env.TW_SECRET_KEY)("deployERC20", () => {
-  it("should deploy ERC20 drop", async () => {
-    const address = await deployERC20Contract({
-      client: TEST_CLIENT,
-      chain: ANVIL_CHAIN,
-      account,
-      type: "DropERC20",
-      params: {
-        name: "TokenDrop",
-        symbol: "NFTD",
-      },
-    });
-    expect(address).toBeDefined();
-    const deployedName = await name({
-      contract: getContract({
-        client: TEST_CLIENT,
-        chain: ANVIL_CHAIN,
-        address,
-      }),
-    });
-    expect(deployedName).toBe("TokenDrop");
-  });
-
-  it("should deploy ERC20 token", async () => {
-    const address = await deployERC20Contract({
-      client: TEST_CLIENT,
-      chain: ANVIL_CHAIN,
-      account,
-      type: "TokenERC20",
-      params: {
-        name: "Token",
-      },
-    });
-    expect(address).toBeDefined();
-    const deployedName = await name({
-      contract: getContract({
-        client: TEST_CLIENT,
-        chain: ANVIL_CHAIN,
-        address,
-      }),
-    });
-    expect(deployedName).toBe("Token");
-  });
-});
