+++ packages/thirdweb/src/extensions/prebuilts/deploy-erc721.test.ts
@@ -11,48 +11,6 @@ const account = TEST_ACCOUNT_B;
 // skip this test suite if there is no secret key available to test with
 // TODO: remove reliance on secret key during unit tests entirely
 describe.runIf(process.env.TW_SECRET_KEY)("deployERC721", () => {
-  it("should deploy ERC721 drop", async () => {
-    const address = await deployERC721Contract({
-      client: TEST_CLIENT,
-      chain: ANVIL_CHAIN,
-      account,
-      type: "DropERC721",
-      params: {
-        name: "NFTDrop",
-      },
-    });
-    expect(address).toBeDefined();
-    const deployedName = await name({
-      contract: getContract({
-        client: TEST_CLIENT,
-        chain: ANVIL_CHAIN,
-        address,
-      }),
-    });
-    expect(deployedName).toBe("NFTDrop");
-  });
-
-  it("should deploy ERC721 token", async () => {
-    const address = await deployERC721Contract({
-      client: TEST_CLIENT,
-      chain: ANVIL_CHAIN,
-      account,
-      type: "TokenERC721",
-      params: {
-        name: "NFTCollection",
-      },
-    });
-    expect(address).toBeDefined();
-    const deployedName = await name({
-      contract: getContract({
-        client: TEST_CLIENT,
-        chain: ANVIL_CHAIN,
-        address,
-      }),
-    });
-    expect(deployedName).toBe("NFTCollection");
-  });
-
   it("should deploy ERC721 open edition", async () => {
     const address = await deployERC721Contract({
       client: TEST_CLIENT,
